Tried colins code

+++ transceiver_drivers.X/transceiver.h
@@ -2,6 +2,6 @@
 
 void transceiver_spi_init(void);
 void rfm_spiwrite(uint32_t c);
-void tft_spiwrite16(unsigned short c);
+void tft_spiwrite16(uint32_t c);
 uint32_t receiver_buffer(void);
 void transmit_buffer(uint32_t data);
